# extension/src/types.gen/models/Preset.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Preset = {
    id?: (number | null);
    name?: string;
    description?: string;
    tags?: Array<string>;
    author?: string;
    content?: string;
};