# extension/src/types.gen/models/SearchFilter.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SearchFilter = {
    id?: (number | null);
    name: string;
    search_query: string;
    author: string;
    tags: Array<string>;
};